Move proxy test to the new class.
