+++ tests/src/test/java/io/streamnative/pulsar/handlers/mqtt/ProxyTest.java
@@ -1,0 +1,42 @@
+package io.streamnative.pulsar.handlers.mqtt;
+
+import io.streamnative.pulsar.handlers.mqtt.base.MQTTTestBase;
+import org.fusesource.mqtt.client.BlockingConnection;
+import org.fusesource.mqtt.client.MQTT;
+import org.fusesource.mqtt.client.Message;
+import org.fusesource.mqtt.client.QoS;
+import org.fusesource.mqtt.client.Topic;
+import org.testng.Assert;
+import org.testng.annotations.AfterClass;
+import org.testng.annotations.BeforeClass;
+import org.testng.annotations.Test;
+
+public class ProxyTest extends MQTTTestBase {
+
+    @BeforeClass
+    @Override
+    public void setup() throws Exception {
+        super.setup();
+    }
+
+    @AfterClass(alwaysRun = true)
+    @Override
+    public void cleanup() throws Exception {
+        super.cleanup();
+    }
+
+    @Test(dataProvider = "mqttTopicNames", timeOut = TIMEOUT)
+    public void testConnectionViaProxy(String topicName) throws Exception {
+        MQTT mqtt = createMQTTProxyClient();
+        BlockingConnection connection = mqtt.blockingConnection();
+        connection.connect();
+        Topic[] topics = { new Topic(topicName, QoS.AT_MOST_ONCE) };
+        connection.subscribe(topics);
+        String message = "Hello MQTT Proxy";
+        connection.publish(topicName, message.getBytes(), QoS.AT_MOST_ONCE, false);
+        Message received = connection.receive();
+        Assert.assertEquals(new String(received.getPayload()), message);
+        received.ack();
+        connection.disconnect();
+    }
+}
